+++ flapper/flapper_test.go
@@ -1,4 +1,4 @@
-package main
+package flapper
 
 import (
 	"testing"
